requirements.txt
asn1crypto==0.24.0
bluepy==1.3.0
certifi==2018.8.24
chardet==3.0.4
cryptography==2.6.1
entrypoints==0.3
idna==2.6
keyring==17.1.1
keyrings.alt==3.1.1
PyBluez==0.23
pycrypto==2.6.1
pygame==2.0.1
PyGObject==3.30.4
python-apt==1.8.4.3
python-dotenv==0.17.1
pyxdg==0.25
requests==2.21.0
SecretStorage==2.3.1
six==1.12.0
ssh-import-id==5.7
urllib3==1.24.1
